+++ qt_helpers.py
@@ -307,7 +307,7 @@ def get_qapp(icon_path=None):
     # Make sure we use high resolution icons with PyQt5 for HDPI
     # displays. TODO: check impact on non-HDPI displays.
     if is_pyqt5():
-        qapp.setAttribute(QtCore.Qt.AA_UseHighDpiPixmaps);
+        qapp.setAttribute(QtCore.Qt.AA_UseHighDpiPixmaps)
 
     return qapp
